完善了人员当前位置的显示，更新了主页图标，但是retrofit用法还待修正

+++ app/src/main/java/com/example/administrator/employeenote/entity/RealLocationData.java
@@ -71,7 +71,7 @@ public class RealLocationData {
 
         private double radius;
 
-        private int speed;
+        private float speed;
 
         private int direction;
 
@@ -111,7 +111,7 @@ public class RealLocationData {
             this.speed = speed;
         }
 
-        public int getSpeed() {
+        public float getSpeed() {
             return this.speed;
         }
